+++ src/mongodb/documents/dish-recipe-document.ts
@@ -2,7 +2,7 @@ import { Document } from 'mongoose';
 import { Language } from '../../common/types';
 import { DishRecipeSections } from '../../modules/dish/dish.types';
 
-export interface RecipeDocument extends Document {
+export interface DishRecipeDocument extends Document {
     readonly language: Language;
     readonly dishId: string;
     readonly sections: DishRecipeSections;