+++ src/mongodb/schemas/dish-comment.schema.ts
@@ -18,3 +18,5 @@ export const DishCommentSchema = new mongoose.Schema({
         required: true
     }
 });
+
+DishCommentSchema.index({ dishId: 1 });